REFACTOR: separate mutation and typedef
